backend/requirements.txt
Django==3.2
django-filter==2.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
Pillow==9.3.0
gunicorn==20.0.4
flake8==4.0.1
psycopg2-binary==2.8.6
